feat(encryption): Make this actually safe
Since, DES encryption is considered outdated by now and not really resistant to brute force attacks, I rewrote the encryption process to use AES-128/192/256 encryption. We use AES-GCM to also verify the integrity of the stored data and rule out invalid keys quickly. This allows us to safely try multiple keys until one works, which enables a new config option that stores previously used keys to try when the current one fails. Starting this version of TBP with a DES encryption key automatically moves it to previousEncryptionKeys and generates a new AES-256 key to be used for new potions.

Besides actual safety, this has the positive side effect that users can now reset their encryption key if it got leaked